+HTML escaping, dup column handling, col mutation
Added HTML escaping of all cell data in the generated
markup string, and column config property `allowHTML` to
bypass it in the case that HTML is expected.

Fixed an issue where if multiple columns referred to the
same key and at elast one had a formatter, that formatter
would affect the rendered content of all like key columns.

Added baseline logic to rerender the tbody if the columns
are changed.  Optimized mutation handling can be relegated
to a plugin.