Add colgroup,col rendering and col width support
Column widths are assigned via CSS on the <col>
node associated with a given column, which is
effectively a min-width.  If content in the column
cells will not fit in the configured width, it will
expand the column.  This is unfortunate native DOM
behavior.

I will create a separate module that enforces column
width-based truncation, but that's not core
functionality.