Updated how Router deals with `popstate` weirdness.
The old implementation wasn't taking the query string into account. The
revamped implementation does, and instead of storing just the `pathname`
it stores the full URL. This way we now can access what the URL used to
be after it has changed, and compare the old URL with the new after
removing the hash fragment.