Deprecate arraylist-add and arraylist-filter
ArrayList has proven more useful as an augmentation class,
relying on custom implmentations of mutation and filter in
subclasses or host classes.  As a low level class, modules
shouldn't be adding things to the prototype that might not
work or be applicable to a subclass.

The standing recommendation is to subclass ArrayList or mix
or augment it onto a class that would benefit from its APIs.
ModelList is a good implementation for mutable ArrayList with
events.

Fixes #2528972
Fixes #2530387
Fixes #2530388