Declare first header column via class
Can't rely on .yui3-datatable-header:first-child because
nested headers can include a rowspan first column, which
makes the first th in the next row (which is not along the
left edge of the table) lose its left border.

Irritatingly, webkit did not behave correctly (hide the border),
so I thought it was a problem common to the other browsers.
Clearly, bias was in play.