WidgetButtons now detaches all event subs it adds to button nodes.
When removing a button, it is not longer destroyed; instead it is
removed from the widget's DOM and any event subscriptions that were
added by the widget are detached.

The `contentUpdate` event is now fired if and only if the button nodes
of a widget's DOM are _actually_ changed.

Add more API docs too.