Force reflow in IE6 before set xscroll table width
The boundingBox width is set before entering _syncXScrollUI
which causes IE to squish any column widths that cause the
overall table width to exceed the bb's dims.  Clearing the
bb width and table width was insufficient to get the table to
expand to natural width (in IE only), so I set the width and
fetch the offsetWidth to trigger an immediate reflow before
proceeding.  The use of the configured table width as the
style width is arbitrary, but possibly a bad call, since it
doesn't take into account borders, and will cause IE to push
out the bb width for a moment.  There's a chance that this could
cause other layout issues or flashing while it updates the UI,
but I'll leave it be for now.