_uiSetWidth needs to account for border width
This unfortunately breaks percentage widths for non-scrolling
tables, since the browser resize is not tracked in core.  The
initial paint should be fine, but resizing the browser won't
resize the table.

I need to move that logic from datatable-scroll to core, but
it'll have to wait for a future sprint :(