Fixed (or at least returned to 3.4.1 state), the sublest regression ever, related to Y.cloning Attribute driven objects with Node values (e.g. Anim), on IE.
Fundamentally, Y.cloning nodes is broken in IE and has been since 3.3.0.

If an object holds multiple references to a Node instance, the first
reference encountered while cloning, will be a damaged clone. Subsequent
references encountered will not clone (and hence be OK).

After the Attribute refactor, the value stored in State was the first
reference encountered, resulting in permanent damage.

Before the refactor, a transient event based reference was the first reference
encountered, and hence was less likely to cause real-world problems.

* Added unit test cases to cover cloning Attribute based objects, with node values
* Tested on IE6, 7, 8, 9.
* Ran some core unit tests (App, Plugin, Widget, Attribute, Base, Node) through Yeti on IE8, Chrome, Safari, FF.
* Spot tested a couple of examples (Widget/App/Attribute).

Wanted to get this in before testfest tomorrow.

Fixes #2531929

Longer term fix for Y.clone(node) on IE is still required.