Fixed regression with WidgetModality's mask not being re-aligned.
For some reason WidgetModality's `scroll` event subscription stopped
working in 3.5.0 which did work in 3.4.1. This changes the event
subscription to explicitly attach to the `window`:

From `Y.on('scroll', ...)` to `Y.one('win').on('scroll', ...)`.

The conditional to check whether this scroll listener should be attached
has been updated to only include iOS < 5 and Android < 3 because the
newer versions of these OSes _actually_ support `position: fixed;`
natively. Whereas the previous versions would pass the feature test but
not actually support `position: fixed;`.

Fixes #2532002