Fixed memory issue with Y.Do not releasing objects cached in Y.Do.objs.
In order to fix this issue cleanly, Y.Do.objs has been deprecated
as of 3.6.0, and will be null.

The cache of AOP'd methods has been moved onto the AOP'd object itself,
so that it gets cleaned up when the object is cleaned up.

Fixes #2529690

* Took the opportunity to fill out Y.Do unit tests.

* Ran Plugin, App, Widget-Stack, Widget-Position, Widget-Position-Align,
Widget-StdMod, Overlay and AutoComplete unit tests (which I believe use Y.Do).

* Confirmed memory issue with Overlay mentioned in the bug is resolved using
Chrome's heap inspector.