Update WidgetButton's `addButton()` method to properly handle `index`.
The `addButton()` method passes its `index` argument to the Array
`splice()` method and was not properly communicating via the
`buttonsChange` event where the new button was at in the section array.
This was happening when the `index` was too large or negative, both of
which the `splice()` method properly handles.

This changes the `index` property on the `buttonsChange` event to be the
index where the new button actually exits. This way the post processing
of the buttons changing can know where the new button is in the array.

Fixes #253219