Duplicate displayColumn code for features (FIXME)
This is a hacky patch to allow features that reference
this._displayColumns to work until they can be migrated
to views or TableView plugins.

This leaves both datatable-base and datatable-table
parsing the display columns from the raw column array.
The TableView should be the source of truth because it
is responsible for what is displayed.