Remove event relay, add property assignment subs
The custom event relay didn't work because
child.on(foo, fire(foo)) executed the after() subs on the
relayed event before the afters or defaultFn on child.

Had to settle for bubbling the table view through DT, allowing
table.on('table:renderHeader', fn).  Not ideal, but...beta? :(

Otherwise, for back compat, the view properties (tableNode, tbodyNode,
theadNode, tfootNode, captionNode, head, body, foot) needed to be
assigned to the DT instance, which for now is done in awkward
event subscriptions to the view events.  captionNode in particular
is now only available in after('table:renderTable') or
after('renderView').  Less than ideal.