Fix Pjax's same-origin link checking process.
Anchor elements in IE behave different than other browsers (surprise!)
in that their `port` property will contain a value even when it's the
default port for the `protocol`. Interestingly enough, IE's `location`
object does _not_ do this, therefore comparing a link's port with the
`location`'s will not work (e.g. "80" !== "").

This adds a method which compares whether or not a link is of the same
origin as the page's current location; it accounts for the default HTTP
and HTTPS ports and normalizes them before the ports are compared.

Fixes #2532456