The example test infrastructure now adds a poll method to Test.Case.prototype.
The method is only added by the example test infrastructure's test-runner.js
impl for now (that is, it's not part of the YUI test module), but it could
be moved to YUI test, with a little more effort to have it integrate more
tightly with wait/resume states and timeouts.

For now, if you're writing example tests, you can use it as follows:

    'my blackbox test which needs to poll' : function() {

       var interval = 10,
           timeout = 10000,
           condition = function() {
             // Return a truthy/falsey result.

             // For example:
             // return Y.one("#waitForMe") !== null
           },
           success = function() {
             Y.Assert.areEqual("StuffYouWantToAssertOnceConditionIsMet", ...);
           },
           failure = function() {
             Y.Assert.fail("#waitForMe never showed up in " + timeout + "ms");
           };

       // failure is optional. Will default to "wait() without resume()" error
       this.poll(condition, interval, timeout, success, failure);

    }

It uses wait/resume under the hood. So poll() starts a  wait(), and we pass
the success/failure callbacks to a resume().

See src/intl/docs/assets/intl-basic-tests.js for usage.