Add `Y.Router.dispatch()` static method which dispatches all routers.
The static `dispatch()` method provides a mechanism to cause all active
router instances to dispatch to their route handlers without needing to
change the URL or fire the `history:change` or `hashchange` event.

This method is used for non-HTML5 routers when `save()` is called but
the location's `hash` isn't actually different, therefore the
`hashchange` event will not be fired. To have parity with how HTML5
routers behave, the static `Y.Router.dispatch()` method is invoked.