Added 'node' as a dependency for widget-buttons-tests
The 'test buttons from another YUI instance' test does a YUI().use('node').

Prior to this change, whether or not that operation was sync or async would
depend on whether or not widget-buttons-tests dependencies had a 'node'
dependency or not.

Prior to gesture-simulate being added as an implicit dependency, 'node's
'node-screen' dependency would be missing from the page for the 2nd use,
and it would be an async operation. The test was coded accordingly with a
wait/resume.

After gesture-simulate was added, 'node-screen' was already on the page,
and the 2nd use would be a synchronous operation, resulting in the wait/resume
construct breaking.

We add 'node' to widget-buttons-tests dependencies, so that the 2nd use
is hopefully now always sync, and the test can be coded accordingly.