Added flag to restore deprecated subscribers/afters properties.
The Y.CustomEvent.MAINTAIN_DEPRECATED_SUBS is false by default and
enabling it, will introduce a performance hit, but will maintain
back compat for the deprecated properties for folks accessing them
directly.

Not quite certain if we'll go the flag route vs. breaking out a
custom-event-base-deprecated module, which will override methods,
but committing it for review.

Went the flag route since it's a small amount of code, and means
we won't have to copy/paste the larger methods which the code
to restore needs to be added to (and hence maintain them until
the properties can be removed).