Add `Y.App.Content`, a `Y.App` class extension.
This extension provides pjax-style content fetching and handling. This
makes it easy to fetch server rendered content for URLs using Ajax. The
HTML content returned from the server will be view-ified and set as the
app's main content, making it seamless to use a mixture of server and
client rendered views.

Closes #2532487