Extended builder logger regex for event-custom-* builds to remove this.log/self.log calls
CE supports a "silent" flag, which allows users to optionally disable log messages in debug
files for as specific type of event. The CustomEvent.prototype.log method was used to
encapsulate the logic for this flag, and hence CustomEvent code used this.log instead of
Y.log directly.

These calls were being left in the min files and resulted in function hop overhead, which
at CEs scale, bubbled up in profiles. NOTE: We never actually logged anything in the min
files, since the Y.log call was stripped. It's just the unneccesary function hop we're trying
to reduce here.