ScrollViewPaginator#scrollToIndex anim duration
ScrollViewPaginator#scrollToIndex takes an animation duration and easing
that aren't having any effect. This is because paginator.set('index',
index) triggers another call to scrollToIndex. This patch fixes that
behavior by passing an "src" option to paginator.set('index', index).

I added tests that check the time spent in the animation (with a margin
of error)