Build deferred and deferred-extras
module entries manually inserted in loader metadata because
running make or shifter cause a lot of file changes, including
'get' and other modules being added to the yui-base.js build file.

Notably, it also added a new definition for 'core' modules that
included yui-log-nodejs, which called require(process....), which
threw an error in the browser env.  Definitely not right. I don't
know what the proper way to rebuild YUI with new module meta is,
but make and shifter definitely aren't it.