Removed IO call, and waitForIOResponse flag from async queue event tests
Looks like it wasn't doing anything with the flag or the call in the test,
and it was breaking when running in the Win 8 App env.

grep'd through code also and nothing under build seems to be doing anything
with the waitForIOResponse flag either.

Tests pass in Chrome/Gecko/IE10 and Win 8 App env.