Normalize Transition durations to ms before updating the node's style.
Instead of relying on browsers to correctly round floating point numbers
it is best to normalize down to ms so we can hopefully be using
integers.

IE10 has an issue with durations under 10ms, when they are represented
in seconds, i.e. `transition-duration: 0.001s`. Normalizing to ms fixed
the issue in IE, and works across YUI's Target Environments.