Add .yeti.json to set project basedir and glob.
The basedir setting makes sure any use of Yeti inside the
project serves your `src` and `build` directories to the
Yeti Hub and therefore the browsers running the tests.

The glob definition is not perfect: the `src/yui` directory
contains files that match the glob but are not always
desired for testing. It is otherwise a suitable default.

http://yeti.cx/docs/v0.2.12/quick-start/#Configuration-file

Note: if you are already using a project-level .yeti.json
to always use a remote Hub, please copy your current
.yeti.json to your home directory. Yeti will consider both
your home directory's settings and the project's configuration
files when setting defaults, but the project's settings take
precedence.