Model validate docs lead to confusion
From the docs it appears as it the callbacks return the error from the validate method and only fall through to the empty callback() if everything is valid. This isn't true however, and both the callback with the error message and the empty callback will be reached. The 'return' is also used in the API docs example: http://yuilibrary.com/yui/docs/api/classes/Model.html#method_validate