Fixed as per Ericf's recommendations except:
opts = opts || {} in AttributeCore

If this were to be moved to each of the public and protected methods,
there would be 4 copies of it in each of set, _set, setAttrs
and _setAttrs for no real purpose at all since opts is not used
in any of them.

The files have been stripped of trailing spacesand
long lines have been shortened except
where a Y.log statement required the line to be all in one
line so it could be effectively deleted in the 'raw' version.
In those cases, /*jshint maxlen:200*/ was added and turned back
to 150 immediately afterwards