Add potential fix for #2532831
This fix patches notifier.fire() to listen for a subsequent click event on the event target, and to preventDefault() on the click. Tested in Chrome, FF, Safari, iOS6, IE6,7,8 and 9. Works everywhere, but IE9 is a little wonky (sometimes mouseups do not fall through to touchends). Not sure what the work around for that is. Need to look into that more.