Fixed bad getSkinName test, showing up as a failure on phantomjs
The class names didn't match the assertions. Moved all the classnames
to be dynamic instead of the static + dynamic, which was getting
confusing to maintain.

Passes on FF, Chrome, Safari, IE, PhantomJS

NOTE: Ideally, want to update the test page to follow the latest
pattern, with test-console, but that's a separate commit.