Added a test to veryfy the bug reported. Added a fix in AttributeObservable.
The problem is that the options are merged into the event
facade and cannot be extricated later on from it to
pass it on to the setter as a separate object.

My solution would be to store it as a separate property
in the same facade so that it is easy to extract.
The added property is prefixed with an underscore to
signal it as for internal use only.