Initial drop of CSSList from the YUI 3 Gallery.
This commit pulls in CSSTable from the YUI 3 Gallery as of gallery-2013.03.27-22-06.
For more granular commits, please refer to the gallery repo available here.
<https://github.com/tilomitra/csslist>

This commit makes some small changes to the `build` and `meta/` files to align
them with YUI. The CSS is up-to-date with ed9f1147.
<https://github.com/tilomitra/csslist/commit/ed9f11470b3732bfe5c8c38cdc597512ef6d0466>

To understand how the CSS works, I recommend opening the manual test page, list.html.