Hacky Base centric solution to bypass event if no listeners
This is a hacky solution to move the hasPotentialSubscribers check
up to the Base layer, so that in can avoid the event system altogether,
if there are no subscribers.

This provides major benefits, and allows us to get to where we'd
like to be with `new Base` [1].

However it's hacky in that:

a. Base knows a lot more about event than it should (especially with
   the fireOnce related properties it's setting).

b. _defInitFn is protected, and may be customized. With the above
   solution, we bypass it if there are no subscribers and go straight to
   the method doing the actual work, but customizations may depend on it
   being called, and being called with the full event facade, instead of
   just the custom payload.

c. _defAttrChangeFn is private, but Model depends on it being called,
   so we need to go through it, as opposed to going straight to
   _setAttrVal(). We also need to provide something which looks like
   an EventFacade instance (in that it has a `stopImmediatePropagation`)
   to support the Model override.

I'm wiling to consider this, if we don't think it's *too* hacky. I'm
thinking of it as a stop gap, until we have more cycles to whittle
down the no listener path further, although I'm not confident we can
get this much of a gain at the end of the day.

Testing
-------

**All** library tests pass through `yogi test --no-cli`.

`[Total]: Passed: 10773 Failed: 0 Total: 10862 (ignored 89)`

Numbers
-------

Base x 71,255
MyBase x 69,479
MyBase with 10 simple value attributes x 33,999
MyBase with 10 simple value attributes - set x 1,172,805
MyModelCore with 2 ad-hoc attrs x 80,430
MyModel with 2 ad-hoc attrs x 47,425

Prior Commit
------------

Base x 48,067
MyBase x 47,107
MyBase with 10 simple value attributes x 25,613
MyBase with 10 simple value atrributes - set x 332,063
MyModel with 2 ad-hoc attrs x 28,722

This amounts to a:

5x increase in `new Base()` over 3.8.1
15x increase in `base.set()` over 3.8.1

for the no-listener use case.

[1] This solution still publishes. There maybe a little more we can
squeeze out here too, but based on what I've looked at we need this
level of publish, just to maintain the event configuration.