Fixed intermittently failing dom, node and widget-parent tests in IE6
Dom/Node tests would fail with a `Node.one() is null` error on load with
a clean cache, and on refresh would sometimes render the console and tests
would pass, and would sometimes not render the console at all.

Issue was that we were creating the YUI instance for test execution in HEAD,
so based on the browser, we could sometimes start executing before the DOM was
fully ready.

Widget-Parent tests would do the same looking for the run button.

Fixed both by moving the script execution to the end of body.

All tests run reliably now in IE6, both manually and through Yeti except for
the Dom's "should find 2 FORMs via form query: Values should be equal"

Ran all tests through chrome as a sanity test, to make sure there
weren't regressions in non-IE browsers.