Attempt to clean up lang xframe tests, to avoid intermittent Yeti w/IE issues
When running the cross frame lang tests on IE, under Yeti, that too,
in a CI context (more load, running all tests etc.) we intermittently see
failures for these tests.

I haven't been able to repro them manually, either running them manually,
or through Yeti locally.

My gut is that we're trying to do stuff with the iframe when it's not
fully ready. In my old life, I remember hitting scenarios where even
though onload fired in the iframe, and the parent could get to
contentWindow, it wouldn't be completely robust when accesing contentWindow
properties.

The fix aims to revert responsibilities. Have the frame call the parent
when ready, as opposed to the parent check something on the frame.

In addition to the attempted fix the commit also does some cleanup:

1). It breaks out the xframe tests into a separate suite, which is not
loaded at all in nodejs (instead of ignoring the whole set of tests
individually)

2). It uses YUITest's async:init support to set up the iframe, so that
everything is encapsulated in the lang-xframe-test.js file, and we
don't need to repeat stuff across the various html file flavors.

NOTE: I didn't find any docs for the async:init stuff in the landing guide,
so had to look through the code - but it seems to be consciously/formally
supported.