Fixed issues due to Y.DOM being undefined when event-base was attached in rare use cases
Fixes #2533242

The `YDOM = Y.DOM` alias we added in 3.7.3 would end up being undefined
in certain custom module dependency use cases, which would result in
`nodelist.on()` failing to attach listners.

In order to bring out the issue, all the following need to apply:

a) Custom module with node-base dependency
b) node-base repeated in use statement
c) Module dependencies for custom modules not provided to YUI()
   instance (not available to loader).

NOTE: This is not the root fix. The root fix is in loader somewhere, but
given the catch22 complexity between node-base and event-base, and the rareness
of the use case, this seemed like the better way to go.

Confirmed that the unit test added passes with this commit, and fails starting
with 3.7.3.