s/EVAL_TOKEN/eval/ and rm supporting script
YUICompressor was unable to minify the json-parse code because it contained
`eval`. So I had to insert a placeholder EVAL_TOKEN which to allow minification,
then use a post-minify script to replace EVAL_TOKEN with `eval`.

This appears not to be necessary any more, and caused a problem with the 3.10.0
CDN deployed files.