Remove Y.Get requirement for tests
The tests were flaky and prone to false-positives due to the dependence on
Y.Get, but these aren't Y.Get tests.

For each test, Y.Get.js is stubbed with a function that calls an indicated
callback from the passed config and/or the test case instance. This allowed
for the removal of asynchronicity in most cases, which really speeds up the
test suite as well.

Also added jsonp-tests to jsonp-url unit test page.