Preserve base jsonp _format if {callback} found
Fixes #700

The _format method override wasn't testing the input url for the {callback}
token *in isolation* and doing a simple replacement if found. It was testing
for the presence of "callback={callback}" instead. That meant the default
behavior changed for 'http://foo.bar?otherparam={callback}' when jsonp-url was
use()d.