Build for AttributeObservable set with opts.
All tests pass through yogi test --no-cli (aside from currently failing
throttle test).

NOTE: This commit in and of itself has a performance hit for the
set(attr, val, opts) case, compared to the prior commit. However,
we're expanding the underlying event argument parsing optimization,
to optimize the path for setting with opts.

Look for this in a following commit. Numbers are in the pull req:
https://github.com/yui/yui3/pull/722