Refactored tests to avoid explicit timeout waits where possible, and stop polling on teardown.
Moved the rest of the tests to a separate test case, and bumped up the timeouts,
even more so for IE, which has been flaky.

I believe the issue was that in some cases (with a 200ms explicit timeout) we were timing
out before the 50ms polling time came around on sluggish machines/browsers.