Big redesign to support row editor in future
To support row editing in the future, cell editors have changed.
There is no longer inline and popup cell editors, just editors.
A `popup` attribute says whether it is a popup or inline.
To convert an inline into popup, the celleditor-overlay can be
plugged in (not a formal plugin) to any editor.
Hopefully, this will help in making possible for the cell editors
to be packed into a row editor just as well.
All the editors in celleditor-popup have been merged along with
those in celleditor-inline are now in celleditors.js