Hacks to make get load tests pass under CI, n/w delays.
2 issues:

a) The purge test was get'ing 30 files. If each had a 0-1s echoecho delay,
   the test was likely to wait(timeout) before it finished.

b) For IE, get currently times out at 3s by default for css requests, since
   onerror is not supported on links.

   With the 0-1s echoecho delay and sluggish CI infrastructure, it had the
   potential to timeout on CSS tests (hitting onFailure, instead of onSuccess).

The hacks:

a) Use a hardcoded delay of 0 for the purge test (not much value in load testing it)
b) Add a large timeout to the CSS requests.

I upped the delay in load.html, to verify the fixes outside of CI.

**NOTE**: These are not production quality changes. Just wanted to capture my hacking
at the root issue as a starting point.

For example, we may not want to add the hight timeout for all browsers, or for the non-load
scenario, just to avoid polluting the tests.