Finished re-arhitecture
The inline and popup editors are now more alike.
Basically, any inline editor can be made a popup by setting
the `popup` attribute to true (which is in fact the case with
many of the popup editors).
The popup editors now are hosted in the Overlay instance in
celleditor-overlay which simply wraps the basic inline editor
in an overlay.
The API docs have been corrected and several CSS style sheets
changed to reflect these changes.