Fixed issue where 'late' fireOnce subscribers would not get a facade, if there were no subscribers during the actual fire.
this.firedWith is updated by reference, during the fire processing,
when a facade is created, and this didn't happen (for valid performance
reasons) if there were no subscribers.

The fix updates this.firedWith with a facade instance, if absent, before
notifying subscribers.

Added unit tests for a bunch of fireOnce scenarios and all library tests
pass.