Added _emptyValues to easily define what is to be considered empty
Instead of having the empty values hardcoded, they are now in an
array of values _emptyValues.
There is now a single criteria to either apply the string formatter
or the emptyCellValue, depending on whether the value to be shown
is in _emptyValues or not.
Test added to check it out.