Starting adding row editing
Added:
* `rowEditor` attr to tell whether the actions set open the row
or cell editor.
* `rowEditorOptions` to allow configuration of row editor.
* `_editRow` prop tells if what's going on is row or cell editing.
* `_rowEditorOverlay` prop holds a reference to the row editor overlay.
* `openRowEditor` method, equivalent to `openCellEditor` for rows.

`_editorTd` is now `_editorNode` since it can hold a TD or a TR.
`_columnEditors` has been dropped in favor of holding the editors.
in property `_editorInstance` of the columns definitions themselves.

Now it uses `_columnMap` to loop over columns, which is always flat
hence easier than using the raw column defs.
In fact, the whole thing failed with nested columns, now it works.