Reverted saving editor instances on column defs
Due to performance reasons, it is not a good idea to store
editor instances in the column defs because the whole info
is deep-cloned, and it takes ages!

Checking the result of firing the cancel event of cell editors
to allow developers to add "Are you sure?" warnings and propagate
the refusal to close.

`hideCellEditor` in now `hideEditor` to let it hide either
kind of editor.

Dropped module configuration info from unit test html files.