Row Editor is now working
Only manually tested, row editing is working.
Also:
References to column key (`colKey` and others similarly named) were
changed to colName and such, since that is what they actually are.
Added `recordKey` for the column key since that is the reference to the
field.
`roweditor-overlay` was renamed simply `roweditor`.
Cell editor's `saveEditor` has a `_quiet` argument to prevent it from
firing the `save` event when part of the row editor.
Dropped `getColumnByTd` and `getColumnNameByTd` since the first is
expected to come from datatable `getColumn` and the second becomes
trivial.
Dropped `getCellEditors` because, since they are shared, it couldn't
really return all the editors at once.
Edited values are not saved if the column definition does not have a
`key` property (as in calculated fields).

Changed the tests and css files to reflect the changes above.