Added documentation-only DataTable.Column class
In order to have some place to put column definitions,
I decided to create a fake class DataTable.Column which
doesn't really have any code but holds the definition
of the properties in the columns definitions.

The column properties have been spread over the files that
actually use them, though the ones in `core.js` should be
distributed into others (like `formatter` in BodyView or
`label` in HeaderView) but they are all listed as belonging
to the base in the user guide.