Changed `columns` into `displayCols`
To make it clearly different from the whole tree of column defs,
I called this flat array `displayCols` but only in the local variables,
the public interfaces are the same.
Fixed the API docs of `refreshRow` which incorrectly stated that the
third argument wasan array of columns while it really is an array of
names.  Also changed the way to look for the columns since the
className might not be built out of the `key`